[skip ci] Create ServiceResource with all the other refactoring that entails.
