+++ dockstore-webservice/src/main/java/io/dockstore/webservice/helpers/GitHubHelper.java
@@ -48,12 +48,17 @@ import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
 import io.dockstore.webservice.CustomWebApplicationException;
 import io.dockstore.webservice.DockstoreWebserviceApplication;
+import io.dockstore.webservice.core.Token;
+import io.dockstore.webservice.core.User;
+import io.dockstore.webservice.jdbi.TokenDAO;
+import io.dockstore.webservice.jdbi.UserDAO;
 import okhttp3.Request;
 import org.apache.commons.io.FileUtils;
 import org.apache.http.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import static io.dockstore.webservice.Constants.LAMBDA_FAILURE;
 import static io.dockstore.webservice.resources.TokenResource.HTTP_TRANSPORT;
 import static io.dockstore.webservice.resources.TokenResource.JSON_FACTORY;
 
@@ -286,4 +291,37 @@ public final class GitHubHelper {
         return repos;
     }
 
+    /**
+     * Based on GitHub username, find the corresponding user
+     * @param tokenDAO
+     * @param userDAO
+     * @param username GitHub username
+     * @param allowFail If true, throw a failure if user cannot be found
+     * @return user with given GitHub username
+     */
+    public static User findUserByGitHubUsername(TokenDAO tokenDAO, UserDAO userDAO, String username, boolean allowFail) {
+        // Find user by github name
+        Token userGitHubToken = tokenDAO.findTokenByGitHubUsername(username);
+        if (userGitHubToken == null) {
+            String msg = "No user with GitHub username " + username + " exists on Dockstore.";
+            LOG.info(msg);
+            if (allowFail) {
+                throw new CustomWebApplicationException(msg, LAMBDA_FAILURE);
+            } else {
+                return null;
+            }
+        }
+
+        // Get user object for github token
+        User sendingUser = userDAO.findById(userGitHubToken.getUserId());
+        if (sendingUser == null) {
+            String msg = "No user with GitHub username " + username + " exists on Dockstore.";
+            LOG.info(msg);
+            if (allowFail) {
+                throw new CustomWebApplicationException(msg, LAMBDA_FAILURE);
+            }
+        }
+
+        return sendingUser;
+    }
 }
